Move test utilities into tests directory
Closes #201.
