+++ tests/test_talreader.py
@@ -2,14 +2,13 @@ from ptsa.data.readers import TalReader
 from ptsa.data.readers.index import JsonIndexReader
 import numpy as np
 import os.path as osp
-from ptsa.test import utils
-import pytest
+from tests import utils
 
 
 @utils.skip_without_rhino
 def test_talreader_on_database():
-    old_reader = TalReader(filename=osp.join(utils.get_rhino_root(),'data/eeg/R1234D/tal/R1234D_talLocs_database_bipol.mat'))
-    jr = JsonIndexReader(osp.join(utils.get_rhino_root(),'protocols/r1.json'))
+    old_reader = TalReader(filename=osp.join(utils.get_rhino_root(), 'data/eeg/R1234D/tal/R1234D_talLocs_database_bipol.mat'))
+    jr = JsonIndexReader(osp.join(utils.get_rhino_root(), 'protocols/r1.json'))
     new_reader = TalReader(filename=jr.get_value('pairs',subject='R1234D',experiment='FR1'))
     assert (new_reader.get_bipolar_pairs() == old_reader.get_bipolar_pairs()).all()
     assert (new_reader.get_monopolar_channels() == old_reader.get_monopolar_channels()).all()
